Replace Skyline, remove articles
I opted to remove the unsupported Skyline altogether, and replace it with the supported fork.
As for the articles, for now I'd prefer to keep the repo focused on tools - e.g. software or datasets. If in the future I have a substantial collection of useful articles I might add a section for those as well. Thanks for the PR!